refactor: move longitudinal_groove to core/detection and update imports
