+++ src/core/detection/longitudinal_groove.py
